fix: Checks call state before checking async/sync state on exit
This change moves the check for isCalled to the top of the exit()
function. Previously, the synchronous check was being checked first,
causing a console error when exit() performed the process.exit() at the
end of the cleanup window.

Includes changes to fixture-async which allow us to control the type of
exit we use when running the fixture, as well as an inclusion of the
failure test.

Fixes #21