feat: Moves `minimumWait` to `wait` and updates to LTS only
There's been ongoing discussion in the original PR and in #32 about
naming the "wait" value. At registration time, it's the minimum amount
of time a function waits, but at `gracefulExit`, it's the largest
minimum wait value. To eliminate this confusion, the option was renamed
to "wait" and the documentation was updated to reflect this.

Because this is a breaking change, we're also co-introducing the LTS
changes. They can be split off into a separate PR if desired, but it
did not add subtantial complexity to this PR.

BREAKING CHANGES:
* `minimumWait` is now `wait`
* Package requires Node 18+ (currently supported LTS)